// src/enums/notification-type.enum.ts
export enum NotificationType {
  CUSTOMER_CREATED = "customer_created",
  PAYMENT_SUBMITTED = "payment_submitted",
  PAYMENT_REVIEWED = "payment_reviewed",
  PAYMENT_CLARIFICATION_REQUESTED = "payment_clarification_requested",
  PAYMENT_CLARIFICATION_RESPONDED = "payment_clarification_responded",
  TASK_ASSIGNED = "task_assigned",
  TASK_SUBMITTED = "task_submitted",
  TASK_REVIEWED = "task_reviewed",
  DESIGNER_APPLICATION = "designer_application",
  DESIGNER_ASSIGNED = "designer_assigned",
  DESIGNER_PHASE_SUBMITTED = "designer_phase_submitted",
  DESIGNER_PHASE_REVIEWED = "designer_phase_reviewed",
  DESIGNER_PAUSED = "designer_paused",
  DESIGNER_RESUMED = "designer_resumed",
  DESIGNER_FINAL_EVALUATED = "designer_final_evaluated",
  QS_TASK_ASSIGNED = "qs_task_assigned",
  QS_EVALUATION_SUBMITTED = "qs_evaluation_submitted",
  QS_EVALUATION_REVIEWED = "qs_evaluation_reviewed",
  DATA_COLLECTOR_SUBMITTED = "data_collector_submitted",
  DATA_COLLECTOR_REVIEWED = "data_collector_reviewed",
  SITE_ENGINEER_SUBMITTED = "site_engineer_submitted",
  SITE_ENGINEER_REVIEWED = "site_engineer_reviewed",
  USER_CREATED = "user_created",
  USER_UPDATED = "user_updated",
  USER_DELETED = "user_deleted",
  SYSTEM_ANNOUNCEMENT = "system_announcement",
}